+++ src/components/atoms/Timeline/hooks.ts
@@ -1,13 +1,27 @@
-import {
+import React, {
   MouseEvent,
   MouseEventHandler,
   useCallback,
   useEffect,
+  useMemo,
   useState,
   WheelEventHandler,
 } from "react";
 
-import { BORDER_WIDTH, GAP_HORIZONTAL, MAX_ZOOM_RATIO, PADDING_HORIZONTAL } from "./constants";
+import {
+  BORDER_WIDTH,
+  DAY_SECS,
+  EPOCH_SEC,
+  SCALE_ZOOM_INTERVAL,
+  GAP_HORIZONTAL,
+  MAX_ZOOM_RATIO,
+  PADDING_HORIZONTAL,
+  SCALE_INTERVAL,
+  MINUTES_SEC,
+  HOURS_SECS,
+  NORMAL_SCALE_WIDTH,
+  STRONG_SCALE_WIDTH,
+} from "./constants";
 import { TimeEventHandler, Range } from "./types";
 
 const convertPositionToTime = (e: MouseEvent, { start, end }: Range, gapHorizontal: number) => {
@@ -25,18 +39,22 @@ const convertPositionToTime = (e: MouseEvent, { start, end }: Range, gapHorizont
 
 type InteractionOption = {
   range: Range;
+  gapHorizontal: number;
+  zoom: number;
+  setZoom: React.Dispatch<React.SetStateAction<number>>;
   onClick?: TimeEventHandler;
   onDrag?: TimeEventHandler;
 };
 
-export const useTimelineInteraction = ({
+const useTimelineInteraction = ({
   range: { start, end },
+  gapHorizontal,
+  zoom,
+  setZoom,
   onClick,
   onDrag,
 }: InteractionOption) => {
   const [isMouseDown, setIsMouseDown] = useState(false);
-  const [zoom, setZoom] = useState(1);
-  const gapHorizontal = GAP_HORIZONTAL * (zoom - Math.trunc(zoom) + 1);
   const handleOnMouseDown = useCallback(() => {
     setIsMouseDown(true);
   }, []);
@@ -94,16 +112,102 @@ export const useTimelineInteraction = ({
 
       setZoom(() => Math.min(Math.max(1, zoom + deltaY * -0.01), MAX_ZOOM_RATIO));
     },
-    [zoom],
+    [zoom, setZoom],
   );
 
   return {
-    zoom,
-    events: {
-      onMouseDown: handleOnMouseDown,
-      onMouseMove: handleOnMouseMove,
-      onClick: handleOnClick,
-      onWheel: handleOnWheel,
-    },
+    onMouseDown: handleOnMouseDown,
+    onMouseMove: handleOnMouseMove,
+    onClick: handleOnClick,
+    onWheel: handleOnWheel,
+  };
+};
+
+const getRange = (range?: { [K in keyof Range]?: Range[K] }): Range => {
+  const { start, end } = range || {};
+  if (start !== undefined && end !== undefined) {
+    return { start, end };
+  }
+
+  if (start !== undefined && end === undefined) {
+    return {
+      start,
+      end: start + DAY_SECS,
+    };
+  }
+
+  if (start === undefined && end !== undefined) {
+    return {
+      start: Math.max(end - DAY_SECS, 0),
+      end,
+    };
+  }
+
+  const defaultStart = Date.now();
+
+  return {
+    start: defaultStart,
+    end: defaultStart + DAY_SECS * EPOCH_SEC,
+  };
+};
+
+type Option = {
+  currentTime: number;
+  range?: { [K in keyof Range]?: Range[K] };
+  onClick?: TimeEventHandler;
+  onDrag?: TimeEventHandler;
+};
+
+export const useTimeline = ({ currentTime, range: _range, onClick, onDrag }: Option) => {
+  const [zoom, setZoom] = useState(1);
+  const range = useMemo(() => {
+    const range = getRange(_range);
+    if (process.env.NODE_ENV !== "production") {
+      if (range.start > range.end) {
+        throw new Error("Out of range error. `range.start` should be less than `range.end`");
+      }
+    }
+    return range;
+  }, [_range]);
+  const { start, end } = range;
+  const startDate = useMemo(() => new Date(start), [start]);
+  const gapHorizontal = GAP_HORIZONTAL * (zoom - Math.trunc(zoom) + 1);
+  const scaleInterval = Math.max(
+    SCALE_INTERVAL - Math.trunc(zoom - 1) * SCALE_ZOOM_INTERVAL * MINUTES_SEC,
+    MINUTES_SEC,
+  );
+  const strongScaleHours = MAX_ZOOM_RATIO - Math.trunc(zoom) + 1;
+  const epochDiff = end - start;
+
+  // convert epoch diff to minutes.
+  const scaleCount = Math.trunc(epochDiff / EPOCH_SEC / scaleInterval);
+
+  // convert HOURS_SECS to minutes.
+  const hoursCount = Math.trunc(HOURS_SECS / scaleInterval);
+
+  // Convert scale count to pixel.
+  const currentPosition = useMemo(() => {
+    const diff = Math.min((currentTime - start) / EPOCH_SEC / scaleInterval, scaleCount);
+    const strongScaleCount = diff / (hoursCount * strongScaleHours);
+    return Math.max(
+      diff * gapHorizontal +
+        (diff - strongScaleCount) * NORMAL_SCALE_WIDTH +
+        strongScaleCount * STRONG_SCALE_WIDTH +
+        STRONG_SCALE_WIDTH / 2,
+      0,
+    );
+  }, [currentTime, start, scaleCount, hoursCount, gapHorizontal, scaleInterval, strongScaleHours]);
+
+  const events = useTimelineInteraction({ range, zoom, setZoom, gapHorizontal, onClick, onDrag });
+
+  return {
+    startDate,
+    scaleCount,
+    hoursCount,
+    gapHorizontal,
+    scaleInterval,
+    strongScaleHours,
+    currentPosition,
+    events,
   };
 };
